pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "get messages from channel in telegram"
authors = [
    {name = "yairbenshi",email = "yairb@alterya.co"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "telethon (>=1.42.0,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "google-cloud-storage (>=3.10.0,<4.0.0)",
    "google (>=3.0.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
